requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
astor==0.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses==0.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
fake-useragent==2.0.3
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
jwt==1.3.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
multidict==6.1.0
Naked==0.1.32
narwhals==1.25.2
Nuitka==2.6
numpy==2.2.2
opencv-python==4.11.0.86
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
pynput==1.7.7
pyobf2==1.3.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python_minifier==2.11.3
pytz==2025.1
pywin32==308
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
selenium==4.28.1
setuptools==75.8.0
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tldextract==5.1.3
tomlkit==0.11.8
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
WMI==1.5.1
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
